Sub 1
Did not submit the first attempt as it couldn't even pass the sample
tests. Read up on the solution in the editorial because I couldn't
figure out the solution. Turned out to be that you could rotate each
cell independent of the others in the quadrant and that's what I was
overlooking.